requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile
#
aniso8601==9.0.1
    # via
    #   -r requirements.in
    #   flask-restful
anyio==3.6.2
    # via
    #   -r requirements.in
    #   httpcore
apscheduler==3.6.3
    # via
    #   -r requirements.in
    #   flask-apscheduler
    #   python-telegram-bot
beautifulsoup4==4.12.0
    # via -r requirements.in
black==23.1.0
    # via -r requirements.in
build==0.10.0
    # via
    #   -r requirements.in
    #   pip-tools
cachetools==4.2.2
    # via
    #   -r requirements.in
    #   google-auth
    #   python-telegram-bot
certifi==2022.12.7
    # via
    #   -r requirements.in
    #   httpcore
    #   httpx
    #   python-telegram-bot
    #   requests
cfgv==3.3.1
    # via
    #   -r requirements.in
    #   pre-commit
charset-normalizer==3.1.0
    # via
    #   -r requirements.in
    #   requests
click==8.1.3
    # via
    #   -r requirements.in
    #   black
    #   flask
    #   pip-tools
distlib==0.3.6
    # via
    #   -r requirements.in
    #   virtualenv
et-xmlfile==1.1.0
    # via
    #   -r requirements.in
    #   openpyxl
filelock==3.10.6
    # via
    #   -r requirements.in
    #   virtualenv
flake8==6.0.0
    # via -r requirements.in
flask==2.2.3
    # via
    #   -r requirements.in
    #   flask-apscheduler
    #   flask-crontab
    #   flask-restful
    #   flask-sqlalchemy
flask-apscheduler==1.12.4
    # via -r requirements.in
flask-crontab==0.1.2
    # via -r requirements.in
flask-restful==0.3.9
    # via -r requirements.in
flask-sqlalchemy==3.0.3
    # via -r requirements.in
google-api-core==2.11.0
    # via
    #   -r requirements.in
    #   google-api-python-client
google-api-python-client==2.82.0
    # via
    #   -r requirements.in
    #   gsheets
google-auth==2.16.3
    # via
    #   -r requirements.in
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-auth-oauthlib
    #   gspread
google-auth-httplib2==0.1.0
    # via
    #   -r requirements.in
    #   google-api-python-client
google-auth-oauthlib==1.0.0
    # via
    #   -r requirements.in
    #   gspread
googleapis-common-protos==1.59.0
    # via
    #   -r requirements.in
    #   google-api-core
greenlet==2.0.2
    # via
    #   -r requirements.in
    #   sqlalchemy
gsheets==0.6.1
    # via -r requirements.in
gspread==5.7.2
    # via -r requirements.in
h11==0.14.0
    # via
    #   -r requirements.in
    #   httpcore
httpcore==0.16.3
    # via
    #   -r requirements.in
    #   httpx
httplib2==0.22.0
    # via
    #   -r requirements.in
    #   google-api-python-client
    #   google-auth-httplib2
    #   oauth2client
httpx==0.23.3
    # via -r requirements.in
identify==2.5.22
    # via
    #   -r requirements.in
    #   pre-commit
idna==3.4
    # via
    #   -r requirements.in
    #   anyio
    #   requests
importlib-metadata==6.1.0
    # via
    #   -r requirements.in
    #   flask
itsdangerous==2.1.2
    # via
    #   -r requirements.in
    #   flask
jinja2==3.1.2
    # via
    #   -r requirements.in
    #   flask
lxml==4.9.2
    # via -r requirements.in
markupsafe==2.1.2
    # via
    #   -r requirements.in
    #   jinja2
    #   werkzeug
mccabe==0.7.0
    # via
    #   -r requirements.in
    #   flake8
mypy-extensions==1.0.0
    # via
    #   -r requirements.in
    #   black
nodeenv==1.7.0
    # via
    #   -r requirements.in
    #   pre-commit
numpy==1.24.2
    # via
    #   -r requirements.in
    #   pandas
oauth2client==4.1.3
    # via
    #   -r requirements.in
    #   gsheets
oauthlib==3.2.2
    # via
    #   -r requirements.in
    #   requests-oauthlib
openpyxl==3.1.2
    # via -r requirements.in
packaging==23.0
    # via
    #   -r requirements.in
    #   black
    #   build
pandas==1.5.3
    # via -r requirements.in
pathspec==0.11.1
    # via
    #   -r requirements.in
    #   black
pip-tools==6.12.3
    # via -r requirements.in
platformdirs==3.2.0
    # via
    #   -r requirements.in
    #   black
    #   virtualenv
pre-commit==3.2.1
    # via -r requirements.in
protobuf==4.22.1
    # via
    #   -r requirements.in
    #   google-api-core
    #   googleapis-common-protos
psycopg2-binary==2.9.5
    # via -r requirements.in
pyasn1==0.4.8
    # via
    #   -r requirements.in
    #   oauth2client
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via
    #   -r requirements.in
    #   google-auth
    #   oauth2client
pycodestyle==2.10.0
    # via
    #   -r requirements.in
    #   flake8
pyflakes==3.0.1
    # via
    #   -r requirements.in
    #   flake8
pyparsing==3.0.9
    # via
    #   -r requirements.in
    #   httplib2
pyproject-hooks==1.0.0
    # via
    #   -r requirements.in
    #   build
python-dateutil==2.8.2
    # via
    #   -r requirements.in
    #   flask-apscheduler
    #   pandas
python-dotenv==1.0.0
    # via -r requirements.in
python-telegram-bot==13.7
    # via -r requirements.in
pytz==2023.2
    # via
    #   -r requirements.in
    #   apscheduler
    #   flask-restful
    #   pandas
    #   python-telegram-bot
pytz-deprecation-shim==0.1.0.post0
    # via
    #   -r requirements.in
    #   tzlocal
pyyaml==6.0
    # via
    #   -r requirements.in
    #   pre-commit
requests==2.28.2
    # via
    #   -r requirements.in
    #   google-api-core
    #   requests-oauthlib
requests-oauthlib==1.3.1
    # via
    #   -r requirements.in
    #   google-auth-oauthlib
rfc3986[idna2008]==1.5.0
    # via
    #   -r requirements.in
    #   httpx
rsa==4.9
    # via
    #   -r requirements.in
    #   google-auth
    #   oauth2client
six==1.16.0
    # via
    #   -r requirements.in
    #   apscheduler
    #   flask-restful
    #   google-auth
    #   google-auth-httplib2
    #   oauth2client
    #   python-dateutil
sniffio==1.3.0
    # via
    #   -r requirements.in
    #   anyio
    #   httpcore
    #   httpx
soupsieve==2.4
    # via
    #   -r requirements.in
    #   beautifulsoup4
sqlalchemy==2.0.7
    # via
    #   -r requirements.in
    #   flask-sqlalchemy
tomli==2.0.1
    # via
    #   -r requirements.in
    #   black
    #   build
    #   pyproject-hooks
tornado==6.2
    # via
    #   -r requirements.in
    #   python-telegram-bot
typing-extensions==4.5.0
    # via
    #   -r requirements.in
    #   black
    #   sqlalchemy
tzdata==2023.2
    # via
    #   -r requirements.in
    #   pytz-deprecation-shim
tzlocal==4.3
    # via
    #   -r requirements.in
    #   apscheduler
uritemplate==4.1.1
    # via
    #   -r requirements.in
    #   google-api-python-client
urllib3==1.26.15
    # via
    #   -r requirements.in
    #   requests
virtualenv==20.21.0
    # via
    #   -r requirements.in
    #   pre-commit
werkzeug==2.2.3
    # via
    #   -r requirements.in
    #   flask
wheel==0.40.0
    # via pip-tools
zipp==3.15.0
    # via
    #   -r requirements.in
    #   importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
